requirements.txt
Flask==2.0.1
requests==2.26.0
bcrypt==4.1.2
grpcio==1.62.1
grpcio-tools==1.62.1